+++ src/components/TodoComponents/Todo.css
@@ -1,7 +1,24 @@
 
 /* todo-list and items */
-.todo-list-container{
+.todo-app-container .header-input{
+  width: 80%;
+  text-align: center;
+  height: 40px;
+  margin-top: 10px;
+  margin-bottom: 15px;
+  font-weight: bold;
+  font-size: 1.8rem;
+  border-radius: 5px;
+}
+
+.todo-app-container .todo-title {
   margin-top: 20px;
+  font-weight: bold;
+  font-size: 2rem;
+}
+
+.todo-list-container{
+  margin-top: 10px;
   width: 100%;
   display: flex;
   flex-direction: column;
